Install new packages each time app cont. comes up
This is useful when we have installed new NPM packages in our local
(host filesystem) and want them to be available inside Docker container
as well. An alternative approach is to mount host node_modules as a
volume in container, but that will override container's own node_modules
folder with host's. Keeping things independent allows for cleaner and
easier troubleshooting of installed/missing packages.