Switch to "page sections" for better authoring
Page Section is our new custom post type dedicated to composing a page
using components (other post types). Each page section is attached to
a UN Page (unpage taxonomy), which gives a way to associate a component
layout with a route/page.

The problem with using unpage directly with components is two-folds:

- unpage taxonomy needs to be attached to be each component's post type

- it's impossible to custom order components on a page*, so defining
layouts such as Hero, ContentWithPreview (1), Testimonials,
ContentWithPreview (2) is not possible

* though it's possible to order components of the same post type (using
a WP plugin such as Post Types Order)

Using page sections solves both the above problems. This comes at a
cost of an additional network call, but is well worth the authoring
flexibility.